1st stage
the front end is almost there
however, it needs to look good.
more features to the front end needs to be added like the background video

nav bar should be collapsible,

first trial for the server, but still not working. However, it looks to good and will work properly ASAP.

contact us section needs to be modified.